add claim and promote tool actions and update submit action to automatically set tool owner
