+++ apps/web/src/components/web/tools/tool-filters.tsx
@@ -4,12 +4,12 @@ import { XIcon } from "lucide-react"
 import { useCallback, useEffect } from "react"
 import type { HTMLAttributes } from "react"
 import { useServerAction } from "zsa-react"
+import { findFilterOptions } from "~/actions/filters"
+import { Badge } from "~/components/common/badge"
 import { Stack } from "~/components/common/stack"
 import { ToolRefinement } from "~/components/web/tools/tool-refinement"
-import { Badge } from "~/components/common/badge"
 import { searchConfig } from "~/config/search"
 import { useToolFilters } from "~/contexts/tool-filter-context"
-import { findFilterOptions } from "~/server/web/tools/actions"
 import type { FilterType } from "~/types/search"
 import { cx } from "~/utils/cva"
